Everything has been tweaked and tested, it all works!
Worked on the interaction between the player and and Ai's and the Ai's on each other.

If the player eats a rage fruit and hits an Ai while in the patrol state, then it will go lose health and switch to its flee state. After the flee timer ahs run out the Ai will go back to the patrol state.